remove deprecated non-integer input to randrange()
Replace float division with integer division in `create_user()` to
ensure integer input to `randint()`, which internally calls
`randrange()`. Non-integer inputs to `randrange()` has been
deprecated in Python 3.10, and will be removed in a future version.

This change also addresses the corresponding python warnings that
would appear in the client code that uses the `create_user()` method.